+++ packages/feathers/index.d.ts
@@ -77,7 +77,7 @@ declare namespace createApplication {
          * A read only property with the name of the service method (one of find, get,
          * create, update, patch, remove).
          */
-        readonly method: string;
+        readonly method: 'find' | 'get' | 'create' | 'update' | 'patch' | 'remove';
         /**
          * A writeable property that contains the service method parameters (including
          * params.query).
@@ -119,21 +119,21 @@ declare namespace createApplication {
         readonly type: 'before' | 'after' | 'error';
     }
 
-    interface HookMap {
-        all: Hook | Hook[];
-        find: Hook | Hook[];
-        get: Hook | Hook[];
-        create: Hook | Hook[];
-        update: Hook | Hook[];
-        patch: Hook | Hook[];
-        remove: Hook | Hook[];
+    interface HookMap<T = any> {
+        all: Hook<T> | Hook<T>[];
+        find: Hook<T> | Hook<T>[];
+        get: Hook<T> | Hook<T>[];
+        create: Hook<T> | Hook<T>[];
+        update: Hook<T> | Hook<T>[];
+        patch: Hook<T> | Hook<T>[];
+        remove: Hook<T> | Hook<T>[];
     }
 
-    interface HooksObject {
-        before: Partial<HookMap> | Hook | Hook[];
-        after: Partial<HookMap> | Hook | Hook[];
-        error: Partial<HookMap> | Hook | Hook[];
-        finally?: Partial<HookMap> | Hook | Hook[];
+    interface HooksObject<T = any> {
+        before: Partial<HookMap<T>> | Hook<T> | Hook<T>[];
+        after: Partial<HookMap<T>> | Hook<T> | Hook<T>[];
+        error: Partial<HookMap<T>> | Hook<T> | Hook<T>[];
+        finally?: Partial<HookMap<T>> | Hook<T> | Hook<T>[];
     }
 
     interface ServiceMethods<T> {
@@ -143,7 +143,7 @@ declare namespace createApplication {
 
         get (id: Id, params?: Params): Promise<T>;
 
-        create (data: Partial<T> | Array<Partial<T>>, params?: Params): Promise<T | T[]>;
+        create (data: Partial<T> | Partial<T>[], params?: Params): Promise<T | T[]>;
 
         update (id: NullableId, data: T, params?: Params): Promise<T | T[]>;
 
@@ -159,7 +159,7 @@ declare namespace createApplication {
     interface ServiceOverloads<T> {
         create? (data: Partial<T>, params?: Params): Promise<T>;
 
-        create? (data: Array<Partial<T>>, params?: Params): Promise<T[]>;
+        create? (data: Partial<T>[], params?: Params): Promise<T[]>;
 
         update? (id: Id, data: T, params?: Params): Promise<T>;