+++ packages/rest-client/test/index.test.js
@@ -1,6 +1,7 @@
 const fetch = require('node-fetch');
 const feathers = require('@feathersjs/feathers');
 const rest = require('../lib/index');
+const { FetchClient } = require('../lib/index');
 const assert = require('assert');
 
 const init = require('../lib');
@@ -92,6 +93,52 @@ describe('REST client tests', function () {
       return service.patch();
     }).catch(error => {
       assert.strictEqual(error.message, `An id must be provided to the 'todos.patch' method`);
+    });
+  });
+
+  it('uses a custom client', () => {
+    const app = feathers();
+    class MyFetchClient extends FetchClient {
+      find () {
+        return Promise.resolve({
+          connection: this.connection,
+          base: this.base,
+          message: 'Custom fetch client'
+        });
+      }
+    }
+
+    app.configure(rest('http://localhost:8889').fetch(fetch, {}, MyFetchClient));
+
+    return app.service('messages').find().then(data => {
+      assert.deepStrictEqual(data, {
+          connection: fetch,
+          base: 'http://localhost:8889/messages',
+          message: 'Custom fetch client'
+        });
+    });
+  });
+
+  it('uses a custom client as second arg', () => {
+    const app = feathers();
+    class MyFetchClient extends FetchClient {
+      find () {
+        return Promise.resolve({
+          connection: this.connection,
+          base: this.base,
+          message: 'Custom fetch client'
+        });
+      }
+    }
+
+    app.configure(rest('http://localhost:8889').fetch(fetch, MyFetchClient));
+
+    return app.service('messages').find().then(data => {
+      assert.deepStrictEqual(data, {
+          connection: fetch,
+          base: 'http://localhost:8889/messages',
+          message: 'Custom fetch client'
+        });
     });
   });
 });